refact: rename directories to lowercase
